stable point functions and packaging
