--- .env.test
@@ -1,5 +1,0 @@
-# region
-AWS_REGION=us-east-1
-
-# api url
-API_URL=https://www.skiagenda.com/test/
