+++ service/packages/client/spec/support/jasmine.json
