+++ service/packages/holidays/holidaysList.js
@@ -1,0 +1,4 @@
+const holidays = require('./holidays')
+exports.handler = (event, context, callback) => {
+    return holidays.list(event, context, callback);
+}
