+++ service/packages/holidays/holidaysRefresh.js
@@ -1,0 +1,4 @@
+const holidays = require('./holidays')
+exports.handler = (event, context, callback) => {
+    return holidays.refresh(event, context, callback);
+}
