+++ service/packages/reservations/deleteReservation.js
@@ -1,0 +1,2 @@
+const reservations = require('./reservations');
+exports.handler = reservations.delete;
