+++ service/packages/reservations/listReservations.js
@@ -1,0 +1,2 @@
+const reservations = require('./reservations');
+exports.handler = reservations.list;
