--- services/lerna.json
@@ -1,8 +1,0 @@
-{
-  "packages": [
-    "packages/*"
-  ],
-  "useWorkspaces": true,
-  "version": "0.0.2",
-  "concurrency": "1"
-}
