--- services/package.json
@@ -1,11 +1,0 @@
-{
-  "workspaces": [
-    "packages/*"
-  ],
-  "devDependencies": {
-    "lerna": "^6.1.0"
-  },
-  "name": "skiagenda-services",
-  "version": "0.0.2",
-  "private": true
-}
