--- services/packages/schoolholidays-db/package.json
@@ -1,23 +1,0 @@
-{
-  "devDependencies": {
-    "jasmine": "^4.5.0",
-    "rimraf": "^3.0.2",
-    "rollup": "^2.60.2"
-  },
-  "dependencies": {
-    "aws-sdk": "^2.566.0",
-    "node-ical": "^0.15.3"
-  },
-  "bundleDependencies": true,
-  "scripts": {
-    "build": "rimraf dist && rollup --config",
-    "test": "jasmine",
-    "pack": "echo 'no pack!'"
-  },
-  "main": "dist/index.js",
-  "name": "schoolholidays-db",
-  "version": "0.0.2",
-  "author": "mcampora",
-  "license": "Apache-2.0",
-  "private": true
-}
