--- services/packages/schoolholidays-db/rollup.config.js
@@ -1,17 +1,0 @@
-//import typescript from 'rollup-plugin-typescript2';
-
-export default {
-    input: ["src/index.js"],
-    output: [
-        {
-            dir: "dist",
-            entryFileNames: "[name].js",
-            format: "cjs",
-            exports: "named"
-        }
-    ],
-    //plugins: [
-    //    typescript(),
-    //],
-    //external: ["react"]
-};
