--- services/packages/schoolholidays/package.json
@@ -1,23 +1,0 @@
-{
-  "devDependencies": {
-    "jasmine": "^4.5.0",
-    "rimraf": "^3.0.2",
-    "rollup": "^2.60.2"
-  },
-  "dependencies": {
-    "schoolholidays-db": "*",
-    "utils": "*"
-  },
-  "bundleDependencies": true,
-  "scripts": {
-    "build": "rimraf dist && rollup --config",
-    "test": "jasmine",
-    "pack": "npm install && mkdir -p ../../build && npm pack --pack-destination ../../build"
-  },
-  "name": "schoolholidays",
-  "version": "0.0.2",
-  "main": "dist/index.js",
-  "author": "mcampora",
-  "license": "Apache-2.0",
-  "private": true
-}
