--- services/packages/schoolholidays/spec/support/jasmine.json
@@ -1,13 +1,0 @@
-{
-  "spec_dir": "spec",
-  "spec_files": [
-    "**/*[sS]pec.js"
-  ],
-  "helpers": [
-    "helpers/**/*.js"
-  ],
-  "stopSpecOnExpectationFailure": true,
-  "oneFailurePerSpec": true,
-  "random": false,
-  "hideDisabled": true
-}
