--- services/packages/utils/package.json
@@ -1,20 +1,0 @@
-{
-  "devDependencies": {
-    "jasmine": "^4.5.0",
-    "rimraf": "^3.0.2",
-    "rollup": "^2.60.2"
-  },
-  "dependencies": {
-  },
-  "scripts": {
-    "build": "rimraf dist && rollup --config",
-    "test": "echo 'no tests!'",
-    "pack": "echo 'no pack!'"
-  },
-  "name": "utils",
-  "main": "dist/index.js",
-  "version": "0.0.2",
-  "author": "mcampora",
-  "license": "Apache-2.0",
-  "private": true
-}
