--- services/packages/utils/src/index.js
@@ -1,13 +1,0 @@
-exports.errorResponse = (errorMessage, awsRequestId, callback) => {
-    callback(null, {
-      statusCode: 500,
-      body: JSON.stringify({
-        Error: errorMessage,
-        Reference: awsRequestId,
-      }),
-      headers: {
-        'Access-Control-Allow-Origin': '*',
-      },
-    });
-}
-
