--- services/scripts/test.sh
@@ -1,5 +1,0 @@
-pushd .
-    cd ../tests
-    npm install
-    jasmine
-popd
