--- services/src.old/reservations-db/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "reservations-db",
-  "version": "0.0.1",
-  "description": "Helper on top of the reservations table",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "mcampora",
-  "license": "Apache-2.0",
-  "dependencies": {
-  }
-}
