--- services/src.old/reservations/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "reservations",
-  "version": "0.0.1",
-  "description": "skiagenda lambda functions",
-  "main": "listReservations.js",
-  "dependencies": {
-    "reservations-db": "file:../reservations-db",
-    "utils": "file:../utils",
-    "ical": "^0.6.0"
-  },
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "mcampora",
-  "license": "Apache-2.0"
-}
